if libAIO can not be loaded, failed explicitely with the OS' name, arch & version

+++ tests/src/org/jboss/messaging/tests/integration/core/asyncio/impl/MultiThreadWriteNativeTest.java
@@ -113,6 +113,14 @@ public class MultiThreadWriteNativeTest extends TestCase
    protected void setUp() throws Exception
    {
        super.setUp();
+       if (!AsynchronousFileImpl.isLoaded())
+       {
+          fail(String.format("libAIO is not loaded on %s %s %s", 
+                System.getProperty("os.name"), 
+                System.getProperty("os.arch"), 
+                System.getProperty("os.version")));
+       }
+
        File file = new File(FILE_NAME);
        file.delete();
